Added value_omitted_from_data method to DynamicArrayWidget.
While testing, I noticed that removing all items from a DynamicArrayWidget and submitting the form would not update the ArrayField. I discovered that similar widgets (including SpitArrayWidget from django.contrib.postgres) need to implement the value_omitted_from_data method. This is usually defined as always returning False. SpitArrayWidget, CheckboxInput, SelectMultiple and CheckboxSelectMultiple all function in this way.

The reason this is important is that when a ModelForm attempts to construct an object instance (see the construct_instance function), it will ignore any field absent from POST data. This appears to be the case for the DynamicArrayWidget.

By defining this method as always False, we force the construct_instance function to always evaluate this field when it is absent (and in our case using the default value from the field)